Change type of n_args in init method from short to int.

+++ insts/jg/PAN/PAN.h
@@ -6,7 +6,7 @@ class PAN : public Instrument {
 public:
    PAN();
    virtual ~PAN();
-   int init(float *, short);
+   int init(float *, int);
    int run();
 };
